broker: abstract out publishing of client data
Make the publishing of client data for route brokers abstract by
changing the kernel client one to remove the dependency on a netlink
message header and for the dataplane client calling a function
pointers, which is populated by default by the previous
implementation. This will allow the producer to provide its own
implementations which don't rely on netlink messages.

VRVDR-50720