broker: provide an object-broker API that doesn't depend on netlink
Provide a more abstract broker API that doesn't use netlink and
instead allows the client to provide implementations of functions that
is flexible enough to use other presentation formats. The API is also
designed to avoid the assumption that there are only two clients of
specific types, although the implementation retains that assumption
for the time being.

Re-implement on the old netlink-specific APIs on top of the new APIs
for backwards compatibility.

VRVDR-50720