broker: advertise protobuf message format to dataplane
Allow the use of multiple presentation formats for data over the same
channel to the dataplane, depending on the broker producer by allowing
the producer to pass in a value corresponding to the format it
supports. This value is opaque to the object broker, although the
route-broker advertises netlink for compatibility reasons, and is
otherwise private between the producer and the consumer.

This allows a system to perform route updates via Google Protocol
Buffers, whilst also allowing a system using FRR to continue to use
netlink.

VRVDR-50720