daemon: call route_broker_show() iff debug is on
brokerd previously called route_broker_show() for each FPM/netlink
message it received. Although route_broker_show(), as called by
brokerd, is a no-op when debugging is not turned on (via --debug
argument) it still iterates the queued broker objects.

This causes a severe drop in performance when brokerd is processing
a large number of routes with the dataplane taking much longer to
converge than expected.

For example, from a cold start of brokerd on DANOS 2009, installing
50,000 unique host routes in the dataplane took ~13 seconds with
these changes. Previously the same operation took ~73 seconds.

DAN-215