daemon: swap RTN_UNSPEC for RTN_UNICAST on rt delete
RTM_DELROUTE messages we receive from FRR always have rtm_type set
to RTN_UNSPEC. This is an issue since those messages are discarded
by the DANOS dataplane.

Therefore on receipt of an RTM_DELROUTE message with rtm_type set to
RTN_UNSPEC, we change it to RTN_UNICAST before passing the message to
the broker code.

Previously we patched this change into the version of FRR used in
DANOS. However doing it within brokerd gives the flexibility to run
vanilla upstream versions of FRR on DANOS. Furthermore this change
also works with both the new "dplane_fpm_nl" FPM implementation, as
well as the existing "fpm" implementation. The previous patch would
require re-work to have an effect when using "dplane_fpm_nl".

VRVDR-53684